+++ NestNet.Sample/SampleApp/Modules/MyModules/Dtos/MyModuleCreateDto.cs
@@ -8,7 +8,7 @@ namespace SampleApp.Modules.MyModules.Dtos
     public class MyModuleCreateDto
     {
 		public required String Name { get; set; }
-		public required Int32 Age { get; set; }
+		public required Int64 Age { get; set; }
 		public String? Email { get; set; }
 		public required String MyVirtualField { get; set; }
     }
