+++ NestNet.Sample/SampleApp/Modules/MyModules/Dtos/MyModuleResultDto.cs
@@ -7,9 +7,9 @@ namespace SampleApp.Modules.MyModules.Dtos
     /// </summary>
     public class MyModuleResultDto
     {
-		public required Int32 MyModuleId { get; set; }
+		public required Int64 MyModuleId { get; set; }
 		public required String Name { get; set; }
-		public required Int32 Age { get; set; }
+		public required Int64 Age { get; set; }
 		public String? Email { get; set; }
 		public required DateTime CreatedAt { get; set; }
 		public required DateTime UpdatedAt { get; set; }
